+++ spec/Nidup/Architool/Application/Refactor/CreateBoundedContextsHandlerSpec.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace spec\Nidup\Architool\Application\BoundedContext;
+namespace spec\Nidup\Architool\Application\Refactor;
 
-use Nidup\Architool\Application\BoundedContext\CreateBoundedContexts;
+use Nidup\Architool\Application\Refactor\CreateBoundedContexts;
 use Nidup\Architool\Domain\Model\BoundedContext;
 use Nidup\Architool\Domain\BoundedContextRepository;
 use PhpSpec\ObjectBehavior;
