+++ src/Nidup/Architool/Application/Project/Pim/CommunityProject.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Nidup\Architool\Application\Project\Pim;
 
-use Nidup\Architool\Application\BoundedContext\CreateBoundedContexts;
+use Nidup\Architool\Application\Refactor\CreateBoundedContexts;
 use Nidup\Architool\Application\Project\Project;
 
 class CommunityProject implements Project
