+++ src/Nidup/Architool/Application/Project/Project.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Nidup\Architool\Application\Project;
 
-use Nidup\Architool\Application\BoundedContext\CreateBoundedContexts;
+use Nidup\Architool\Application\Refactor\CreateBoundedContexts;
 
 interface Project
 {
